+++ src/feeds/hashtags.ts
@@ -7,9 +7,52 @@ import Toot from "../api/objects/toot";
 import UserData from "../api/user_data";
 import { bracketed } from "../helpers/string_helpers";
 import { config } from "../config";
-import { MastodonTag, CacheKey } from "../types";
+import { buildTagNames, countTags } from "../api/objects/tag";
+import { MastodonTag, CacheKey, StringNumberDict } from "../types";
+import { sortKeysByValue, truncateToConfiguredLength } from "../helpers/collection_helpers";
 import { traceLog } from "../helpers/log_helpers";
-import { truncateToConfiguredLength } from "../helpers/collection_helpers";
+import { mastodon } from "masto";
+
+
+export async function getFavouritedTagToots(): Promise<Toot[]> {
+    const logPrefix = bracketed("getFavouritedHashtagToots()");
+    const favouritedTagCounts = countTags(await MastoApi.instance.getFavouritedToots());
+    const followedTags = buildTagNames(await MastoApi.instance.getFollowedTags());
+    const participatedTags = await UserData.getUserParticipatedTags();
+
+    let favouritedNotParticipatedTagCounts = Object.entries(favouritedTagCounts).reduce(
+        (acc, [tagName, count]) => {
+            if (config.trending.tags.invalidTrendingTags.includes(tagName)) {
+                return acc;
+            }
+
+            // TODO: filter out tags with low particiaption with a heuristic based on favourited tag counts
+            if (!followedTags[tagName] && (participatedTags[tagName]?.numToots || 0) <= 2) {
+                acc[tagName] = count;
+            }
+
+            return acc;
+        },
+        {} as StringNumberDict
+    );
+
+    let tagNames = sortKeysByValue(favouritedNotParticipatedTagCounts).slice(0, config.favouritedTags.numTags);
+
+    tagNames.forEach(tagName => {
+        traceLog(`${logPrefix} Favourited not followed/participated tag: ${tagName} (${favouritedTagCounts[tagName]} times)`);
+    });
+
+    // TODO: this sucks
+    const tags = tagNames.map(name => {
+        return {name: name, url: MastoApi.instance.tagUrl(name), history: []} as mastodon.v1.Tag;
+    });
+
+    return await MastoApi.instance.getCacheableToots(
+        async () => await MastoApi.instance.getStatusesForTags(tags, config.favouritedTags.numTootsPerTag),
+        CacheKey.FAVOURITED_HASHTAG_TOOTS,
+        config.favouritedTags.maxToots,
+    );
+};
 
 
 // Get recent toots from hashtags the user has participated in frequently