+++ src/types.ts
@@ -49,6 +49,7 @@ export type WeightName = ScoreName | NonScoreWeightName;
 export enum CacheKey {
     BLOCKED_ACCOUNTS = 'BlockedAccounts',
     FAVOURITED_TOOTS = 'FavouritedToots',
+    FAVOURITED_HASHTAG_TOOTS = 'FavouritedHashtagToots',
     FEDIVERSE_POPULAR_SERVERS = 'FediversePopularServers',
     FEDIVERSE_TRENDING_TAGS = 'FediverseTrendingTags',
     FEDIVERSE_TRENDING_LINKS = 'FediverseTrendingLinks',